refactor package structure and class name
