+++ server2/src/main/kotlin/com/junktion/api/v1/models/image/ImageRepository.kt
@@ -1,0 +1,5 @@
+package com.junktion.api.v1.models.image
+
+interface ImageRepository {
+	fun register()
+}
