+++ server2/src/main/kotlin/com/junktion/api/v1/models/image/ImageService.kt
@@ -1,0 +1,5 @@
+package com.junktion.api.v1.models.image
+
+interface ImageService {
+	fun upload(): String
+}
