--- server2/src/main/kotlin/com/junktion/api/v1/models/repositories/ImageRepository.kt
@@ -1,5 +1,0 @@
-package com.junktion.api.v1.models.repositories
-
-interface ImageRepository {
-	fun register()
-}
