+++ server2/src/main/kotlin/com/junktion/api/v1/services/ImageServiceImpl.kt
@@ -1,16 +1,17 @@
 package com.junktion.api.v1.services
 
-import com.junktion.api.v1.models.repositories.ImageRepository
+import com.junktion.api.v1.models.image.ImageRepository
+import com.junktion.api.v1.models.image.ImageService
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.stereotype.Service
 
 @Service
-class ImageService(
+class ImageServiceImpl(
         private val imageRepository: ImageRepository
-) {
+): ImageService {
 
     @PreAuthorize("#oauth2.hasScope('default')")
-    fun upload(): String {
+    override fun upload(): String {
         imageRepository.register()
         return "TODO"
     }
